# src/main/resources/static/openapi.yml
openapi: 3.0.3
info:
  title: Rating System API
  description: API documentation for Rating System
  version: 1.0.0
servers:
  - url: ${app.base-url}
    description: Local server
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: ${AUTH_SERVER_URL}/realms/leverx/protocol/openid-connect/auth
          tokenUrl: ${AUTH_SERVER_URL}/realms/leverx/protocol/openid-connect/token
security:
  - oauth2: [ ]
...
paths:
...
  /api/v1/auth:
    post:
      tags:
        - Auth
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                username:
                  type: string
                email:
                  type: string
                password:
                  type: string
                confirmPassword:
                  type: string
                role:
                  type: string
      responses:
        '201':
          description: User registered successfully
  /api/v1/auth/confirmation:
    post:
      tags:
        - Auth
      summary: Activate user account
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Account activated successfully
  /api/v1/auth/forgot-password:
    post:
      tags:
        - Auth
      summary: Send reset password email
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Password reset email sent
  /api/v1/auth/new-password:
    post:
      tags:
        - Auth
      summary: Reset user password
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
        - name: token
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                confirmPassword:
                  type: string
      responses:
        '204':
          description: Password reset successfully
...